Add Allow and Restrict to PluginConfigOptionValue
This commit allows Allow and Restrict fields to PluginConfigOptionValue.
The purpose of these fields is to facilitate a uniform method of simple
option validation through equality testing. The equality testing is
provided via go-cmp. The zero values of Allow and Restrict result in the
same behaviour as before they were added, so this change is
backwards-compatible.

Signed-off-by: Eric Chlebek <eric@sensu.io>